Scraping Twitter Data around certain topics
These lines of code allow you to scrape tweets that contain the words Trump & China
The idea was here to compare the general opinion of the public on President Trump by comparing the sentiment score of tweets containing these 2 words before and after a certain date, based on an important statement made by Donald Trump on that given date. 
Furthermore with the use of the N-grams function the commonly appearing singular, duos and trios of words are listed and finally a network of terms has also been created to understand the relationship between the different other words appearing in the tweets besides trump & china (both words are removed as they are not relevant anymore since we have already selected the tweets which are containing these words).